Remove mockito dependency from dynamodb
